refactor: useMoreSheet 훅 추가 및 적용
