+++ src/components/Card/index.ts
@@ -1,1 +1,2 @@
 export { default as GroupingCard } from './GroupingCard.client';
+export { default as CardHeader } from './CardHeader.client';
